pyproject.toml
[tool.poetry]
name = "django-backend"
version = "0.1.0"
description = ""
authors = ["mnogoruk <lionless072@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.5"
djangorestframework = "^3.12.2"
whitenoise = "^5.2.0"
gunicorn = "^20.0.4"
django-heroku = "^0.3.1"
python-dotenv = "^0.15.0"
django-cors-headers = "^3.7.0"
djangorestframework-simplejwt = "^4.6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
